Split target tiflash_vector_features for vector features
