+++ dbms/src/Flash/Disaggregated/tests/gtest_s3_lock_service.cpp
@@ -110,7 +110,7 @@ protected:
 #define CHECK_S3_ENABLED                                                                                          \
     if (!is_s3_test_enabled)                                                                                      \
     {                                                                                                             \
-        const auto * t = ::testing::UnitTest::GetInstance()->current_test_info();                                 \
+        const auto * t = ::testing::UnitTest::GetInstance() -> current_test_info();                               \
         LOG_INFO(log, "{}.{} is skipped because S3ClientFactory is not inited.", t->test_case_name(), t->name()); \
         return;                                                                                                   \
     }
