+++ dbms/src/TiDB/Decode/Vector.cpp
@@ -16,28 +16,9 @@
 #include <IO/WriteBufferFromString.h>
 #include <IO/WriteHelpers.h>
 #include <TiDB/Decode/Vector.h>
+#include <VectorFeatures/simdsimd-internals.h>
 
 #include <compare>
-
-// SIMSIMD is header only. We don't use cmake to make these defines to avoid
-// polluting all compile units.
-
-// Note: Be careful that usearch also includes simsimd with a customized config.
-// Don't include simsimd and usearch at the same time. Otherwise, the effective
-// config depends on the include order.
-#define SIMSIMD_NATIVE_F16 0
-#define SIMSIMD_NATIVE_BF16 0
-#define SIMSIMD_DYNAMIC_DISPATCH 0
-
-// Force enable all target features. We will do our own dynamic dispatch.
-#define SIMSIMD_TARGET_NEON 1
-#define SIMSIMD_TARGET_SVE 0 // Clang13's header does not support enableing SVE for region
-#define SIMSIMD_TARGET_HASWELL 1
-#define SIMSIMD_TARGET_SKYLAKE 1
-#define SIMSIMD_TARGET_ICE 0
-#define SIMSIMD_TARGET_GENOA 0
-#define SIMSIMD_TARGET_SAPPHIRE 0
-#include <simsimd/simsimd.h>
 
 namespace DB
 {
@@ -46,59 +27,6 @@ namespace ErrorCodes
 {
 extern const int BAD_ARGUMENTS;
 } // namespace ErrorCodes
-
-namespace simsimd_details
-{
-
-simsimd_capability_t simd_capabilities()
-{
-    static simsimd_capability_t static_capabilities = simsimd_cap_any_k;
-    if (static_capabilities == simsimd_cap_any_k)
-        static_capabilities = simsimd_capabilities_implementation();
-    return static_capabilities;
-}
-
-simsimd_capability_t actual_capability(simsimd_datatype_t data_type, simsimd_metric_kind_t kind)
-{
-    simsimd_metric_punned_t metric = nullptr;
-    simsimd_capability_t used_capability;
-    simsimd_find_metric_punned(
-        kind,
-        data_type,
-        simsimd_details::simd_capabilities(),
-        simsimd_cap_any_k,
-        &metric,
-        &used_capability);
-
-    return used_capability;
-}
-
-} // namespace simsimd_details
-
-std::vector<std::string> VectorDistanceSIMDFeatures::get()
-{
-    simsimd_capability_t cap_l2 = simsimd_details::actual_capability(simsimd_datatype_f32_k, simsimd_metric_l2sq_k);
-    simsimd_capability_t cap_cos = simsimd_details::actual_capability(simsimd_datatype_f32_k, simsimd_metric_cos_k);
-
-    auto cap_to_string = [](simsimd_capability_t cap) -> String {
-        if (cap & simsimd_cap_neon_k)
-            return "neon";
-        if (cap & simsimd_cap_sve_k)
-            return "sve";
-        if (cap & simsimd_cap_sve2_k)
-            return "sve2";
-        if (cap & simsimd_cap_haswell_k)
-            return "haswell";
-        if (cap & simsimd_cap_skylake_k)
-            return "skylake";
-        return "serial";
-    };
-
-    std::vector<std::string> ret{};
-    ret.push_back("vec.l2=" + cap_to_string(cap_l2));
-    ret.push_back("vec.cos=" + cap_to_string(cap_cos));
-    return ret;
-}
 
 VectorFloat32Ref::VectorFloat32Ref(const Float32 * elements, size_t n)
     : elements(elements)
