+++ libs/libcommon/include/common/logger_useful.h
@@ -45,7 +45,7 @@ inline constexpr size_t getFileNameOffset(const T (&str)[S], size_t i = S - 1)
 }
 
 template <typename T>
-inline constexpr size_t getFileNameOffset(T (&/*str*/)[1])
+inline constexpr size_t getFileNameOffset(T (& /*str*/)[1])
 {
     return 0;
 }
